Added Build.PL command-line option --local-alien
I found that I wanted to hack a bit on Alien::TinyCC, and realized
that the easiest way to do that and see the impact of my changes
would be to shoehorn the already provided local Alien::TinyCC
machinery. Since I already installed the latest Aline::TinyCC, the
Build.PL script would not notice anything and not offer the install,
so I added this option to the Build.PL file to do it form me. :-)