Updated lexical variable interpolation.
Lexical variable interpolation was using preprocessor macros which
sometimes called a function and other times extracted members directly
from my_perl, depending on how Perl was built. In the latter case,
tcc would complain that it didn't know how to extract elements from
a void pointer. The solution is to ensure that the pointer is
properly cast.

Although this commit fixes variable interpolation, other perl.h
macros may fail. I worked on a fix which enabled setting the type of
my_perl to PerlInterpreter, in a way which could become incresingly
robust, but even the first attempt did not work, so I am using the
direct casting method.