Require library linking to go through compiler_options global, simplify libperl
This fixes the static library loading issue with libperl. There is
still a problem: I get a segfault. However, it no longer chokes on
the static library, which is a win. I think this uncovered a bug in
the extended symbol table handling in tcc, so I'll have to go back
to that to figure out what's going on.