Store the results of the check_identifier test
The check_identifier tests take a very long time. Part of me wanted
to paralellize them, but then I'd have to write code that could
adapt to prallel vs non-parallel, or forking vs non-forking perls.
It would be nice, eventually, but for now I'm going to stick with
single-process identifier checking. But to speed up development,
I store the results of the checking in a separate file, and use
those results. This makes testing and revising libperl essentially
instantaneous.