Extract code into three separate SVs instead of the buffer.
I want to make it easy and sensible to add code chunks to the
top and bottom of a cblock. I could keep going with the
preprocessor definitions, but that always seemed like a hack to
me. I now collect pre-declarations and cleanup code in separate
SVs, and the main code body in a third one. These are all
concatenated before compiling.