Clean up C code parser/extractor
The parser/extractor was returning magic integers, so I switched
these to be defined by an enum. I also found a couple of minor
cleanups along the way. I still need to work on line directives
and exceptions during interpolation blocks, but at least it's a
little easier to follow.