Remove cisa!
This commit officially removes the cisa keyword, replacing it with
a mechanism that uses Perl's built-in type annotations. These type
annotations are meant to be used with the fields pragma, but it's
just too useful to leave alone in this context. I also took the
liberty of changing the naming convention for types to be much,
much shorter.

There are still some files with cisa keywords; these will be cleaned
out in the next few commits.