Add two new benchmarks
In preparation for entries for the C::Blocks Advent, I wrote a
couple of new benchmarks for producing random numbers and, then,
using those random numbers to perform random lookups in a large
array. The second of these was meant to test performance that is
being throttled by RAM latency. (It does not give the result I
expected, though.)