Remove cache bias in random access test
It makes sense to have the Inline::C and C::Blocks functions access
the exact same random slots in memory, because it lets me verify
that they are actually doing the same thing. However, the first to
access the data may end up placing the data in the processors cache,
biasing the results in favor of the second. This commit causes the
two implementations to reach for different random locations, leading
to a fairer comparison.