Prevent at least one class of leak-on-exception
... by introducing automatic scope boundary cleanup.
The same effect could be gotten for most of the cleanup
by simply mortalizing some of the SVs, but this way is
more generic by invoking the cleanup function.

The downside is that one has to make sure that anything (other than
exceptions) that might return from the keyword plugin function early
invokes LEAVE. A subsequent patch will refactor that to be asserted via
a simple wrapper function.

Several more such leaks remain.