Build and invoke custom OPs slightly more efficiently
In a nutshell, this changes goes from executing two OPs (CONST puts
function pointer on stack, XOP pops it off and executes it) to a single
OP: We simply shove the function pointer into one of the unused struct
members, op_targ.

The normal purpose of op_targ (targ, short for target, sigh) is to store
an offset into the PADLIST for re-using SVs in the functions lexical
(scratch)pad for passing data between OPs instead of using the stack.
Because all the pushing, popping, and especially repeated SV allocation
comes at a cost.

That's the "t3" in the following example OP tree:

  $ perl -MO=Concise -e 'my $x = $_/2;'
  8  <@> leave[1 ref] vKP/REFC ->(end)
  1     <0> enter ->2
  2     <;> nextstate(main 1 -e:1) v:{ ->3
  7     <2> sassign vKS/2 ->8
  5        <2> divide[t3] sK/2 ->6
  -           <1> ex-rv2sv sK/1 ->4
  3              <#> gvsv[*_] s ->4
  4           <$> const[IV 2] s ->5
  6        <0> padsv[$x:1,2] sRM*/LVINTRO ->7
  -e syntax OK

It's kind of a cool optimization, really. :)

The one gotcha is that perl normally helpfully frees the targeted
scratchpads, so we have to define an OP free hook that sets it to 0 to
prevent funny free's.