Move OP building logic to the custom OP related source file
This slightly refactors it so that the TCC and keyword-type related
logic isn't in it. For now, this seems the cleanest in intent. If a lot
of additional logic was added (eg. OPs owning miscellaneous data that
would have to be cleaned up at global destruction time), that should be
revisited to move that logic into the new source file as well.