Fix keyword scoping implementation
The logic for this is borrowed from PLua: In a nutshell, we install our
keyword plugin handler into the chain of keyword plugins globally and
permanently, and then we use the hints hash system to dynamically modify
whether or not we're actually APPLYING our keyword handler in practice.

Doing this another way seems to be infeasible because you'd otherwise
have to have a way to go through the entire chain of keyword handlers
and plucking out this one in the unimport. Otherwise, this module can
never work with any other keyword plugin (globally).