Restructure parser/extractor code marginally
In a nutshell: "Public" functions are cb_ prefixed. Private ones are
made static. Due to the size of this .c file and the number of private
functions, make sure to declare all functions at the top to avoid any
sort of function order dependencies that add cognitive overhead when
coding.