Expand Types::Struct
It is now possible to significantly tweak a struct declaration. This
leads to greater flexibility and reduces the likelihood of name
collisions. It is now also possible to get the list of elements for
a struct, presumably to be used in creating extended structs with
beginnings that are binary compatible with other structs.