Make PerlAPI included by default
I have yet to find a situation in which I don't at some point want
the Perl API. Having it off by default is silly. This commit
changes things so that it's on, and only excluded if you explicitly
ask for it. This likely will requie a change to some docs, and
definitely some tutorials.