Fix line number counting.
This patch addresses line number counting. The trick in this case was
making sure that both data->end and PL_bufptr were sensible. Up to this
point, data->end was good, but PL_bufptr was sometimes off. In short,
lex_unstuff(loc) tells the buffer that loc is good data. I was not quite
sure if loc was the end of the bad data, or the start of the good data.
With that resolved (start of good data), we can fix all the problems with
buffering and line counting.