Add new keyword cq{}
This commit adds a new keyword cq{}. Its job is to provide a string
with the enclosed C code. It uses the exact same parsing mechanism
as the other C::Blocks keywords, and so behaves exactly like them in
its text extraction and filtering. However, instead of compiling
the result, it produces a Perl string with the contents. This makes
it very easy to produce a small snippet of C code in a Perl string
which has been filtered and which correctly reports the line number
where it was actually written. This should help ensure that when the
compiler encounters an error, the files and line numbers it reports
are much more likely to be correct and helpful. The application of
lexical C::Blocks filters helps keep the effects of filters lexically
scoped, as they should be.

Tests with this keyword revealed a bug in how line numbers were
reported to Perl interpolation blocks. This commit also includes
the fix of that bug.