Make scalars of type C::Blocks::Types::Struct read-only
While thinking over the structure of the basic types, I admitted
that the opaque structs using this type setup were totally
unprotected. I then realized that there is a Perl flag to
protect that, so this commit turns that on for any scalar that is
annotated as a struct.

The write protection only takes effect after the first cblock in
which the scalar is used. (The code that accomplishes this is
generated by c_blocks_init_cleanup.)