Revise cq to try to handle variable interpolation; improve cq docs
I anticipate using cq a lot in SOS. I began looking over that and I
saw that I use double quotes and variable interpolation a lot. I
realized that variable interpolation would really be quite helpful,
and so I decided I needed to revise cq to allow for that. This
commit implements that (requiring quite a bit of new C parsing code)
and also significantly expands the documentation on cq and other
code generation mechanisms.