Make cq's line directives optional
In separate work I began converting bits of SOS to use cq{}. A
mistake made early was almost immediately diagnosable thanks to
the line number handling. However, the line number make for *much*
more verbose text, and it occurred to me that almost certainly
this will lead to additional processing time which is not useful
after code has been debugged. For this reason, I decided to make
line directives optional, and globally controlled. (The changes
to SOS will land in a separate commit.)